feat: add new commands for all the template and data file
