+++ messages/smockit.template.init.md
@@ -1,0 +1,20 @@
+# summary
+
+For creating a test data template.
+
+# description
+
+More information about a command. Don't repeat the summary. 
+
+# flags.name.summary
+
+Description of a flag.
+
+# flags.name.description
+
+More information about a flag. Don't repeat the summary. 
+
+# examples
+
+- '$ sf template init --default'
+
