+++ messages/smockit.template.print.md
@@ -1,0 +1,15 @@
+# summary
+
+This will print the content of the data template in the console.
+
+# flags.templateName.summary
+
+Template Name to print on CLI
+
+# flags.templateName.description
+
+Use `--templateName` or `-t` to specify the name of the data template to be printed. The template must exist in the `data_gen/templates` directory.
+
+# Examples
+
+- `sf template print -t <template-name>`
