+++ src/services/output-formatter.ts
@@ -16,7 +16,7 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { GenericRecord } from '../utils/types.js';
-import DataGenerate from '../commands/data/generate.js';
+import DataGenerate from '../commands/smockit/data/generate.js';
 
 export function saveOutputFileOfJsonAndCsv(
   jsonData: GenericRecord[],
@@ -42,9 +42,8 @@ export function saveOutputFileOfJsonAndCsv(
 
 export function saveCreatedRecordIds(outputFormat: string[], templateName: string): void {
   if (outputFormat.includes('DI') || outputFormat.includes('di')) {
-    const fileName = `${templateName?.replace('.json', '')}_createdRecords_${
-      new Date().toISOString().replace('T', '_').replace(/[:.]/g, '-').split('.')[0]
-    }.json`;
+    const fileName = `${templateName?.replace('.json', '')}_createdRecords_${new Date().toISOString().replace('T', '_').replace(/[:.]/g, '-').split('.')[0]
+      }.json`;
     saveMapToJsonFile('data_gen', fileName);
   }
 }
@@ -79,4 +78,3 @@ export function convertJsonToCsv(jsonData: GenericRecord[]): string {
   const csvRows = data.map((row: Record<string, any>) => fields.map((field: string) => row[field] as string).join(','));
   return [fields.join(','), ...csvRows].join('\n');
 }
-
