+++ test/commands/smockit/data/upload.test.ts
@@ -1,0 +1,40 @@
+import { TestContext } from '@salesforce/core/testSetup';
+import { expect } from 'chai';
+import { stubSfCommandUx } from '@salesforce/sf-plugins-core';
+import SmockitDataUpload from '../../../../src/commands/smockit/data/upload.js';
+
+describe('smockit data upload', () => {
+  const $$ = new TestContext();
+  let sfCommandStubs: ReturnType<typeof stubSfCommandUx>;
+
+  beforeEach(() => {
+    sfCommandStubs = stubSfCommandUx($$.SANDBOX);
+  });
+
+  afterEach(() => {
+    $$.restore();
+  });
+
+  it('runs hello', async () => {
+    await SmockitDataUpload.run([]);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello world');
+  });
+
+  it('runs hello with --json and no provided name', async () => {
+    const result = await SmockitDataUpload.run([]);
+    expect(result.path).to.equal('src/commands/smockit/data/upload.ts');
+  });
+
+  it('runs hello world --name Astro', async () => {
+    await SmockitDataUpload.run(['--name', 'Astro']);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello Astro');
+  });
+});
