+++ test/commands/smockit/template/print.test.ts
@@ -1,0 +1,40 @@
+import { TestContext } from '@salesforce/core/testSetup';
+import { expect } from 'chai';
+import { stubSfCommandUx } from '@salesforce/sf-plugins-core';
+import SmockitTemplatePrint from '../../../../src/commands/smockit/template/print.js';
+
+describe('smockit template print', () => {
+  const $$ = new TestContext();
+  let sfCommandStubs: ReturnType<typeof stubSfCommandUx>;
+
+  beforeEach(() => {
+    sfCommandStubs = stubSfCommandUx($$.SANDBOX);
+  });
+
+  afterEach(() => {
+    $$.restore();
+  });
+
+  it('runs hello', async () => {
+    await SmockitTemplatePrint.run([]);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello world');
+  });
+
+  it('runs hello with --json and no provided name', async () => {
+    const result = await SmockitTemplatePrint.run([]);
+    expect(result.path).to.equal('src/commands/smockit/template/print.ts');
+  });
+
+  it('runs hello world --name Astro', async () => {
+    await SmockitTemplatePrint.run(['--name', 'Astro']);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello Astro');
+  });
+});
