+++ test/commands/smockit/template/remove.test.ts
@@ -1,0 +1,40 @@
+import { TestContext } from '@salesforce/core/testSetup';
+import { expect } from 'chai';
+import { stubSfCommandUx } from '@salesforce/sf-plugins-core';
+import SmockitTemplateRemove from '../../../../src/commands/smockit/template/remove.js';
+
+describe('smockit template remove', () => {
+  const $$ = new TestContext();
+  let sfCommandStubs: ReturnType<typeof stubSfCommandUx>;
+
+  beforeEach(() => {
+    sfCommandStubs = stubSfCommandUx($$.SANDBOX);
+  });
+
+  afterEach(() => {
+    $$.restore();
+  });
+
+  it('runs hello', async () => {
+    await SmockitTemplateRemove.run([]);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello world');
+  });
+
+  it('runs hello with --json and no provided name', async () => {
+    const result = await SmockitTemplateRemove.run([]);
+    expect(result.path).to.equal('src/commands/smockit/template/remove.ts');
+  });
+
+  it('runs hello world --name Astro', async () => {
+    await SmockitTemplateRemove.run(['--name', 'Astro']);
+    const output = sfCommandStubs.log
+      .getCalls()
+      .flatMap((c) => c.args)
+      .join('\n');
+    expect(output).to.include('hello Astro');
+  });
+});
